renamed 'product_types_widget' to 'product_types_combo'
